fix(project selector): Use `slug` rather than `name` to identify projects (#57)
The UI for a project's `name` was pulled from project settings almost a year ago, in this PR: https://github.com/getsentry/sentry/pull/11986.

However, we still fill in the field at project creation time, with a title-kebab-cased version of the platform if no other name is specified, and the name in turn generates the slug. So, for new projects, `name` and `slug` match (save possible case differences).

The problem comes when someone changes a project's slug in the UI (which many folks do, since initially many projects have only the platform as the slug, which is too generic). Because `name` used to be separately controllable from the UI, it made sense to not touch it when updating the slug. But now what that means is that unless you specify a name at project creation time, it will forever be equal to the platform. And since the wizard's project selector currently looks at `name` instead of `slug`, people get confused, especially if they have more than one project with the same platform-as-name `name` value.

So, long story long, this changes the project selector to use `slug` instead, which isn't deprecated, matches the UI, and is guaranteed to be unique within the org.

NB: This is only not a one-liner because prettier had opinions about how things were formatted and changed a bunch of stuff. The only relevant change is in the second red -> green block.

[see PR for links to code that does the various things mentioned above]