fix(nextjs): Don't expose auth token in `sentry.properties` (#128)
The Wizard is currently adding every property required by Sentry CLI in `sentry.properties`, including the auth token (that shouldn't be exposed). This PR fixes this behavior by adding the auth token to `.sentryclirc` (the Sentry CLI config file), which is added to the `.gitignore`. `sentry.properties` can now be safely committed.