ref(nextjs): Check version range rather than minimum version (#143)
Next 12, which has just been released, includes some changes which break the SDK. As a result, for the moment we're not officially supporting it. This fixes the compatibility check which occurs as part of the nextjs wizard to reflect that reality. (It also simplifies the implementation of said check a bit, relying more on the `semver` library and less on manual checks.)