feat(nextjs): Add logic to add `@sentry/nextjs` if it's missing (#219)
This PR adds logic to add the @sentry/nextjs package if it's not installed and we either have a yarn.lock or package-lock.json file. The purpose of this change is to enable a more seamless installation process where the user doesn't require the additional step of installing @sentry/nextjs.

Note that this only works if there is a lock file because we need that to know which package manager to install. IMO this is probably OK. I would imagine most folks would already have that file if they are thinking about adding Sentry. I think it's unlikely that people at Sentry before even installing packages and running whatever they are using locally.

Also note that this means the nextjs installation works slightly differently than the other platforms. In the future, we can back propagate these changes if we find them to be helpful.