fix: Update minimum Node version to Node 14 (#332)
In the v3 major bump, we implicitly raised the minimum node version to Node 14 and in #248 we already switched to Node 14 for building in CI. In `v3.0.0`, the wizard was theoretically still compatible with lower Node versions. In `v3.1.0` when we introduced `clack`-based wizard flows, Node 14 became a strict requirement. However, we missed raising the `engine.node` min version in `package.json` to `>=14.0.0` in the initial 3.0.0 release.

This patch now updates the minimum Node version to 14 and ammends the changelog for 3.0.0 to let folks know.