feat(sourcemaps): Detect SvelteKit and NextJS projects and redirect to dedicated wizards (#341)
Detect SvelteKit and NextJS project and redirect to dedicated wizards

We observed that several users run the source maps wizard in NextJs (or potentially SvelteKit) projects which isn't ideal. With this change we let them know that a better option exists for them and offer to redirect them to the dedicated wizard.