feat(sourcemaps): Add CLI-based flow for Angular (#349)
Add a CLI-based source maps upload configuration flow for Angular.

The reason is that for Angular devs, it might not be apparent that the "best" option for them previously was "None of the above". Some clicked "Webpack" although Angular doesn't offer access to the webpack config by default. Hence this option, similarly to CRA, simply redirects to the CLI flow and adds a custom source maps generation step.