ref(sourcemaps): Always collect source maps wizard telemetry (#348)
Previously, we only collected telemetry when the source maps wizard was directly started with the -i sourcemaps arg. This is probably the majority of cases anyway. However, for the sake of consistency, we should always collect telemetry. Therefore, this patch slightly refactors the location of the withTelemetry call so that we're now always using it.