ref(apple): Move `withTelemetry` into `runAppleWizard` (#353)
Align the apple wizard with the refactor of the source maps wizard in #348 where we moved the `withTelemetry` call into the wizard. The benefit is that we have the call in one place instead of two and that we can potentially check the `telemetryEnabled` flag whenever we need to within the wizard, as it's now part of the wizard options.