feat(sourcemaps): Pre-select auto-detected build tool option (#354)
Add simple auto-detection of installed build tool packages to preselect a tool selection option in the source maps wizard. Detection is performed by checking the users' package.json for specific packages. Either the packages of the build tool/bundler or the framework/library (Angular/CRA). If nothing is found, we fall back to the catch-all "None of the above" option which starts the Sentry CLI flow.