fix: Ensure wizard exits after setup is completed (#360)
When the wizard was started with the `-i` flag, it would sometimes just keep on awaiting a promise after the `run[integration]Wizard` function returned. Seems like we broke the promise chain in #348 or sometime before that. This caused the application to appear hanging and users needing to cancel the wizard manually. 

Affected wizards.
* NextJs
* Sveltekit
* Sourcemaps
* Apple (not tested but I'm fairly certain)

This patch makes sure that the wizard exits by removing our duplicated invokation logic for the `clack`-based wizard. For the moment, we always go through the Shim classes. I'd like to refactor this eventually and get rid of the shims but for now we should fix this to unblock users.