fix(telemetry): Re-enable telemetry collection (#361)
Seems like a check in #334 unintentionally disabled telemetry collection by default. I missed this when reviewing the PR. As a consequence of this and merging in #360, we didn't collect any telemetry data in the last 3.7.0 release. This patch removes the unnecessary check and restores telemetry collection.

Furthermore, this patch also adjusts the welcome message to no longer show the telemetry collection note in the welcome message if it is disabled.