fix: Refactor the wizard version lookup to handle gracefully (#816)
The lookup of the `package.json` of the `@sentry/wizard` can throw when used during development outside of the repository. This is especially relevant developing the apple wizard, which is most likely in a project directory unrelated to the wizard.

The current implementation will crash the wizard, rendering it unusable. Ideally we are hard-baking the version on `yarn build`, but until then the dynamic lookup must not throw an error if the package.json is not found.

It is better to have telemetry without a node version, instead of crashing the wizard.