feat(sveltekit): Add support for SDK setup with `instrumentation.server.ts` (#1077)
Add support for setting up the SDK with SvelteKit's new `instrumentation.server.ts` [instrumentation file](https://svelte.dev/docs/kit/observability), as well as by leveraging the new spans emitted from kit. This requires a SvelteKit version of at least 2.31.0. Of course we want to stay backwards compatible, so this PR adds the following flow:

We check for version from `package.json` (I'm, this is flawed and could mean we identify a version <2.31.0 when in fact it is already used. But then again, if not explicitly set in package.json, we can't rely on >= 2.31.0 being used). If we identify <2.31.0, we recommend updating to the newest version. 

From this point, there are 3 options for the user:
  1. update kit: exits the wizard and lets the user update kit. I don't think we should do this for them.
  2. users are already using >=2.31.0 (i.e. false negative from our end): Continue with new instrumentation file setup
  3. user will upgrade later: Continue with old, hooks-based setup.

If we detect a version >=2.31.0, we don't ask anything but directly continue with the new setup.

For the old setup, nothing changes.

For the new setup, we now apply the following changes:
- enable instrumentation and tracing in `svelte.config.js`
- add `Sentry.init` into `instrumentation.server.mjs`
- add handle and error hooks into `hooks.server.ts`
- add Sentry.init and error hook into `hooks.client.ts`

---------

Co-authored-by: Andrei <168741329+andreiborza@users.noreply.github.com>